Rename masks to azimuthal_masks for clarity
