+++ map_engraver/data/proj/azimuthal_masks.py
